## Why
500_region-full-porting.md Phase 6 の実装。C版watershed.c の高レベルAPI
（per-basin画像抽出・塗りつぶし描画・カラー描画）をRustに移植する。

## What
- `WatershedResult` 構造体追加（per-basin Pix + labeled Pix）
- `watershed_with_basins()` - セグメンテーション結果をbasin別に分解
- `watershed_render_fill()` - 各basinを最小グレー値で塗りつぶした8bit画像
- `watershed_render_colors()` - 各basinを疑似ランダムカラーで着色した32bit RGBA画像
- 境界ピクセルを透明黒ではなく不透明黒 (`0x000000FF`) で明示設定
- `ccbord.rs`: 不要なi32/u32同一型キャストを削除（clippy）

## Impact
`leptonica-region::watershed` モジュールのみ。他crateへの影響なし。

PR: #167

+++ crates/leptonica-region/src/ccbord.rs
@@ -492,8 +492,8 @@ impl ImageBorders {
             .write_all(b"ccba")
             .map_err(|e| RegionError::InvalidParameters(e.to_string()))?;
 
-        let width_bytes = (self.width as u32).to_le_bytes();
-        let height_bytes = (self.height as u32).to_le_bytes();
+        let width_bytes = self.width.to_le_bytes();
+        let height_bytes = self.height.to_le_bytes();
         writer
             .write_all(&width_bytes)
             .map_err(|e| RegionError::InvalidParameters(e.to_string()))?;
@@ -509,10 +509,10 @@ impl ImageBorders {
 
         // Per-component data
         for comp in &self.components {
-            let bx = (comp.bounds.x as i32).to_le_bytes();
-            let by = (comp.bounds.y as i32).to_le_bytes();
-            let bw = (comp.bounds.w as i32).to_le_bytes();
-            let bh = (comp.bounds.h as i32).to_le_bytes();
+            let bx = comp.bounds.x.to_le_bytes();
+            let by = comp.bounds.y.to_le_bytes();
+            let bw = comp.bounds.w.to_le_bytes();
+            let bh = comp.bounds.h.to_le_bytes();
 
             writer
                 .write_all(&bx)
@@ -699,8 +699,8 @@ fn write_border<W: Write>(writer: &mut W, border: &Border) -> RegionResult<()> {
         .map_err(|e| RegionError::InvalidParameters(e.to_string()))?;
 
     // Start point
-    let sx = (border.start.x as i32).to_le_bytes();
-    let sy = (border.start.y as i32).to_le_bytes();
+    let sx = border.start.x.to_le_bytes();
+    let sy = border.start.y.to_le_bytes();
     writer
         .write_all(&sx)
         .map_err(|e| RegionError::InvalidParameters(e.to_string()))?;
@@ -721,8 +721,8 @@ fn write_border<W: Write>(writer: &mut W, border: &Border) -> RegionResult<()> {
 
     // Points
     for p in &border.points {
-        let px = (p.x as i32).to_le_bytes();
-        let py = (p.y as i32).to_le_bytes();
+        let px = p.x.to_le_bytes();
+        let py = p.y.to_le_bytes();
         writer
             .write_all(&px)
             .map_err(|e| RegionError::InvalidParameters(e.to_string()))?;
